Refactors ES events
