+++ module/Althingi/src/Events/DeleteEvent.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Althingi\ServiceEvents;
+namespace Althingi\Events;
 
 use Althingi\Presenters\IndexablePresenter;
 use Althingi\Presenters\IndexablePresenterAwareInterface;
 use Zend\EventManager\Event;
 
-class AddEvent extends Event implements IndexablePresenterAwareInterface
+class DeleteEvent extends Event implements IndexablePresenterAwareInterface
 {
     /** @var  \Althingi\Presenters\IndexablePresenter */
     private $presenter;
 
     public function __construct(IndexablePresenter $presenter)
     {
-        parent::__construct('add', $this);
+        parent::__construct('delete', $this);
         $this->setPresenter($presenter);
     }
