+++ module/Althingi/src/Service/Congressman.php
@@ -6,8 +6,8 @@ use Althingi\Lib\DatabaseAwareInterface;
 use Althingi\Lib\EventsAwareInterface;
 use Althingi\Model\CongressmanValue as CongressmanValueModel;
 use Althingi\Presenters\IndexableCongressmanPresenter;
-use Althingi\ServiceEvents\AddEvent;
-use Althingi\ServiceEvents\UpdateEvent;
+use Althingi\Events\AddEvent;
+use Althingi\Events\UpdateEvent;
 use Althingi\Model\Congressman as CongressmanModel;
 use Althingi\Model\CongressmanAndParty as CongressmanAndPartyModel;
 use Althingi\Model\CongressmanAndCabinet as CongressmanAndCabinetModel;
