+++ module/Althingi/src/Service/Issue.php
@@ -15,8 +15,8 @@ use Althingi\Hydrator\AssemblyStatus as AssemblyStatusHydrator;
 use Althingi\Model\IssueTypeStatus as IssueTypeStatusModel;
 use Althingi\Hydrator\IssueTypeStatus as IssueTypeStatusHydrator;
 use Althingi\Presenters\IndexableIssuePresenter;
-use Althingi\ServiceEvents\AddEvent;
-use Althingi\ServiceEvents\UpdateEvent;
+use Althingi\Events\AddEvent;
+use Althingi\Events\UpdateEvent;
 use Althingi\Hydrator\IssueValue as IssueValueHydrator;
 use Althingi\Model\IssueValue as IssueValueModel;
 use InvalidArgumentException;
