+++ module/Althingi/src/Service/Party.php
@@ -11,8 +11,8 @@ use Althingi\Hydrator\PartyAndElection as PartyAndElectionHydrator;
 use Althingi\Hydrator\PartyAndTime as PartyAndTimeHydrator;
 use Althingi\Lib\DatabaseAwareInterface;
 use Althingi\Presenters\IndexablePartyPresenter;
-use Althingi\ServiceEvents\AddEvent;
-use Althingi\ServiceEvents\UpdateEvent;
+use Althingi\Events\AddEvent;
+use Althingi\Events\UpdateEvent;
 use Zend\EventManager\EventManager;
 use Zend\EventManager\EventManagerInterface;
 use PDO;
