+++ module/Althingi/src/Service/Speech.php
@@ -5,8 +5,8 @@ namespace Althingi\Service;
 use Althingi\Lib\DatabaseAwareInterface;
 use Althingi\Lib\EventsAwareInterface;
 use Althingi\Presenters\IndexableSpeechPresenter;
-use Althingi\ServiceEvents\AddEvent;
-use Althingi\ServiceEvents\UpdateEvent;
+use Althingi\Events\AddEvent;
+use Althingi\Events\UpdateEvent;
 use Althingi\Hydrator\Speech as SpeechHydrator;
 use Althingi\Hydrator\SpeechAndPosition as SpeechAndPositionHydrator;
 use Althingi\Hydrator\DateAndCount as DateAndCountHydrator;
