+++ module/Althingi/tests/Service/PartyTest.php
@@ -3,7 +3,7 @@
 namespace AlthingiTest\Service;
 
 use Althingi\Service\Party;
-use Althingi\ServiceEvents\ServiceEventsListener;
+use Althingi\ElasticSearchActions\ElasticSearchEventsListener;
 use AlthingiTest\DatabaseConnection;
 use AlthingiTest\ElasticBlackHoleClient;
 use PHPUnit\Framework\TestCase;
@@ -129,7 +129,7 @@ class PartyTest extends TestCase
 
     public function testCreate()
     {
-        $serviceEventListener = (new ServiceEventsListener())
+        $serviceEventListener = (new ElasticSearchEventsListener())
             ->setElasticSearchClient(new ElasticBlackHoleClient())
             ->setLogger(new NullLogger());
         $eventManager = new EventManager();
@@ -160,7 +160,7 @@ class PartyTest extends TestCase
 
     public function testSave()
     {
-        $serviceEventListener = (new ServiceEventsListener())
+        $serviceEventListener = (new ElasticSearchEventsListener())
             ->setElasticSearchClient(new ElasticBlackHoleClient())
             ->setLogger(new NullLogger());
         $eventManager = new EventManager();
@@ -191,7 +191,7 @@ class PartyTest extends TestCase
 
     public function testUpdate()
     {
-        $serviceEventListener = (new ServiceEventsListener())
+        $serviceEventListener = (new ElasticSearchEventsListener())
             ->setElasticSearchClient(new ElasticBlackHoleClient())
             ->setLogger(new NullLogger());
         $eventManager = new EventManager();
